+++ R/AllGenerics.R
@@ -32,6 +32,12 @@ setGeneric("export", function(object, ...)
 setGeneric("filterIntensity", function(object, ...)
     standardGeneric("filterIntensity"))
 #' @rdname hidden_aliases
+setGeneric("filterMzRange", function(object, ...)
+    standardGeneric("filterMzRange"))
+#' @rdname hidden_aliases
+setGeneric("filterMzValues", function(object, ...)
+    standardGeneric("filterMzValues"))
+#' @rdname hidden_aliases
 setGeneric("isReadOnly", function(object, ...)
     standardGeneric("isReadOnly"))
 #' @rdname hidden_aliases