+++ R/peaks-functions.R
@@ -92,6 +92,54 @@ NULL
 
 #' @description
 #'
+#' Filter the spectrum keeping only peaks that match the provided m/z value(s).
+#'
+#' @inheritParams .peaks_remove
+#'
+#' @param mz `numeric` with the m/z values to match against. This needs to be
+#'     a sorted `numeric` (has to be checked upstream).
+#'
+#' @param tolerance `numeric` with the tolerance. Can be of length 1 or equal
+#'     length `mz`.
+#'
+#' @param ppm `numeric` with the ppm. Can be of length 1 or equal length `mz`.
+#'
+#' @author Johannes Rainer
+#'
+#' @importFrom MsCoreUtils closest
+#'
+#' @noRd
+.peaks_filter_mz_value <- function(x, spectrumMsLevel, mz = numeric(),
+                                   tolerance = 0, ppm = 20,
+                                   msLevel = spectrumMsLevel, ...) {
+    if (!spectrumMsLevel %in% msLevel || !length(x))
+        return(x)
+    idx <- closest(mz, x[, "mz"], tolerance = tolerance, ppm = ppm,
+                   duplicates = "closest", .check = FALSE)
+    x[idx[!is.na(idx)], , drop = FALSE]
+}
+
+#' @description
+#'
+#' Filter the spectrum keeping only peaks that are within the provided m/z
+#' range.
+#'
+#' @inheritParams .peaks_remove
+#'
+#' @param mz `numeric(2)` with the lower and upper m/z value.
+#'
+#' @author Johannes Rainer
+#'
+#' @noRd
+.peaks_filter_mz_range <- function(x, spectrumMsLevel, mz = numeric(),
+                                   msLevel = spectrumMsLevel, ...) {
+    if (!spectrumMsLevel %in% msLevel || !length(x))
+        return(x)
+    x[between(x[, "mz"], mz), , drop = FALSE]
+}
+
+#' @description
+#'
 #' Bin peaks of a spectrum.
 #'
 #' @inheritParams .peaks_remove